feat: implement GreeMQTTApp class for enhanced device management and asynchronous operations
